requirements.txt
# Xử lý video, audio
moviepy==1.0.3
opencv-python
pydub
numpy
Pillow

# ML / AI frameworks
torch
torchaudio

# Whisper (OpenAI)
git+https://github.com/openai/whisper.git

# Transformers (nếu có dùng cho NLP hoặc xử lý thêm)
transformers

# YOLOv8 (ultralytics)
ultralytics

#khác
tqdm
google-genai
ipywidgets